dict updates
Primarily reshuffling the j/ch/sh conundrum.

basically, previously:
a) we had the letter j pronounced like in "jam"
b) the sound 'sh' in foreign words was assumed to be transliterated with the 's'

however, having just one extremely arbitrary affricate didn't actually make sense, but putting every sh/j/ch/etc sound with the 's' would make transliterating Mandarin a literal fucking nightmare (and it was already difficult), so we added/replaced j with the 'sh' sound, transliterated to a x because x is a weird stupid letter that no one knows how to pronounce anyway. and it's an 'sh' in pinyin anyway.

So I had to go back and fix all the transliterations this fucked up, repick words for all the newly formed duplicates, and just generally fuck with stuff.